pyproject.toml
[project]
name = "road-health"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.29.0",
    "opencv-python>=4.12.0.88",
]
